+++ osidb/models.py
@@ -2831,9 +2831,6 @@ class FlawAcknowledgment(ACLMixin, BugzillaSyncMixin, TrackingMixin):
         Flaw, on_delete=models.CASCADE, related_name="acknowledgments"
     )
 
-    # non operational meta data
-    meta_attr = HStoreField(default=dict)
-
     objects = FlawAcknowledgmentManager()
 
     class Meta:
@@ -2933,9 +2930,6 @@ class FlawReference(ACLMixin, BugzillaSyncMixin, TrackingMixin):
 
     # one flaw can have many references
     flaw = models.ForeignKey(Flaw, on_delete=models.CASCADE, related_name="references")
-
-    # non operational meta data
-    meta_attr = HStoreField(default=dict)
 
     objects = FlawReferenceManager()