+++ osidb/tests/test_flaw.py
@@ -172,7 +172,6 @@ class TestFlaw:
             vuln_1,
             "https://www.openwall.com/link123",
             type=FlawReference.FlawReferenceType.EXTERNAL,
-            meta_attr={"foo": "bar"},
             acl_read=self.acl_read,
             acl_write=self.acl_write,
         )
@@ -188,7 +187,6 @@ class TestFlaw:
             "name",
             "company",
             from_upstream=True,
-            meta_attr={"foo": "bar"},
             acl_read=self.acl_read,
             acl_write=self.acl_write,
         )